feat: add pouch selection

+++ src/modules/pouches/ui/pouch-modal/pouch-item/pouch-item.tsx
@@ -2,28 +2,37 @@ import { PouchEntity, PouchId } from '@pv/core/entities/pouch.entity'
 import { Icon } from '@pv/ui-kit/icon'
 import { useCallback } from 'preact/compat'
 import { Button } from '@pv/ui-kit/button'
-import { ActionWrapper, Container, IconWrapper, TitleWrapper } from './pouch-item-styles'
+import { ActionWrapper, Container, IconWrapper, Row, TitleWrapper } from './pouch-item-styles'
 
 type Props = {
+  isSelected: boolean
   pouch: PouchEntity
-  onRemove: (id: PouchId) => void
+  onRemove?: (id: PouchId) => void
+  onSelect: (id: PouchId) => void
 }
 
-export function PouchItem({ pouch, onRemove }: Props) {
+export function PouchItem({ pouch, onRemove, isSelected, onSelect }: Props) {
   const handleRemove = useCallback(() => {
-    onRemove(pouch.id)
+    onRemove && onRemove(pouch.id)
   }, [onRemove, pouch])
+
+  const handleSelect = useCallback(() => {
+    onSelect(pouch.id)
+  }, [onSelect, pouch.id])
+
   return (
     <Container>
-      <IconWrapper>
-        <Icon iconName="wallet" />
-      </IconWrapper>
-      <TitleWrapper>{pouch.name}</TitleWrapper>
-      <ActionWrapper>
-        <Button variant="flat" onClick={handleRemove}>
-          <Icon iconName="cross" />
-        </Button>
-      </ActionWrapper>
+      <Row onClick={handleSelect}>
+        <IconWrapper>{isSelected && <Icon iconName="wallet" />}</IconWrapper>
+        <TitleWrapper>{pouch.name}</TitleWrapper>
+      </Row>
+      {onRemove && (
+        <ActionWrapper>
+          <Button variant="flat" onClick={handleRemove}>
+            <Icon iconName="cross" />
+          </Button>
+        </ActionWrapper>
+      )}
     </Container>
   )
 }
